update: 파라미터에 imageUrl 필드 추가 및 로직 수정

+++ src/main/java/team/incude/gsmc/v2/domain/evidence/application/usecase/service/UpdateOtherEvidenceByCurrentuserService.java
@@ -8,14 +8,12 @@ import team.incude.gsmc.v2.domain.evidence.application.port.EvidencePersistenceP
 import team.incude.gsmc.v2.domain.evidence.application.port.OtherEvidencePersistencePort;
 import team.incude.gsmc.v2.domain.evidence.application.port.S3Port;
 import team.incude.gsmc.v2.domain.evidence.application.usecase.UpdateOtherEvidenceByCurrentUserUseCase;
-import team.incude.gsmc.v2.domain.evidence.domain.ActivityEvidence;
 import team.incude.gsmc.v2.domain.evidence.domain.Evidence;
 import team.incude.gsmc.v2.domain.evidence.domain.OtherEvidence;
 import team.incude.gsmc.v2.domain.evidence.domain.constant.ReviewStatus;
 import team.incude.gsmc.v2.global.thirdparty.aws.exception.S3UploadFailedException;
 
 import java.io.IOException;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -27,20 +25,31 @@ public class UpdateOtherEvidenceByCurrentuserService implements UpdateOtherEvide
 
     @Override
     @Transactional
-    public void execute(Long evidenceId, MultipartFile file) {
+    public void execute(Long evidenceId, MultipartFile file, String imageUrl) {
         Evidence evidence = evidencePersistencePort.findEvidenceByIdWithLock(evidenceId);
-        deleteFileByEvidenceId(evidenceId);
+        OtherEvidence otherEvidence = otherEvidencePersistencePort.findOtherEvidenceById(evidenceId);
+
+        String fileUri = checkImageUrl(otherEvidence, imageUrl, file);
 
         Evidence newEvidence = createEvidence(evidence);
-        String fileUrl = uploadFile(file);
-        OtherEvidence newOtherEvidence = createOtherEvidence(newEvidence, fileUrl);
-
+        OtherEvidence newOtherEvidence = createOtherEvidence(newEvidence, fileUri);
         otherEvidencePersistencePort.saveOtherEvidence(newOtherEvidence);
     }
 
-    private void deleteFileByEvidenceId(Long evidenceId) {
-        OtherEvidence otherEvidence = otherEvidencePersistencePort.findOtherEvidenceById(evidenceId);
+    private String checkImageUrl(OtherEvidence otherEvidence, String imageUrl, MultipartFile file) {
+        if (imageUrl != null
+                && !imageUrl.isEmpty()
+                && otherEvidence.getFileUri().equals(imageUrl)) {
+            return imageUrl;
+        }
+
         s3Port.deleteFile(otherEvidence.getFileUri());
+
+        if (file != null && !file.isEmpty()){
+            return uploadFile(file);
+        } else {
+            return null;
+        }
     }
 
     private Evidence createEvidence(Evidence evidence) {
